Move geodetic representations to earth.py
